fix(vite-plugin-angular): fix resolution of external templates for builds
